Separate Action-Required Tests from Required tests
Added two buttons, one of which runs required tests and the other that runs all tests.  Action-required tests require you to perform an action (e.g. switching a theme or going to different files).
